refactor(netsuite-tba): update import paths for schemas
