+++ integrations/netsuite-tba/types.ts
@@ -1,6 +1,3 @@
-import { workerData } from 'worker_threads';
-import { lineItemDefaultPropertiesSchema } from '../hubspot/schema.zod';
-
 export interface NSAPI_GetResponse<T> {
     data: T;
 }
@@ -448,323 +445,10 @@ export interface NS_PurchaseOrder {
     tranDate?: string;
     tranId?: string;
 }
-////////////////////
-// vendor bill
-// export interface VendorBill {
-//     vendorBill: string;
-//     account: string;
-//     accountingBookDetail: accountingBookDetail[];
-//     appliedRules: appliedRules[];
-//     approvalStatus: NS_Reference;
-//     availableVendorCredit: number;
-//     balance: number;
-//     billAddr1: string;
-//     billAddr2: string;
-//     billAddr3: string;
-//     billAddress: string;
-//     billAddressList: nsResource;
-//     billAddressee: string;
-//     billAttention: string;
-//     billCity: string;
-//     billCountry: NS_Reference;
-//     billOverride: string;
-//     billPhone: string;
-//     billState: string;
-//     billZip: string;
-//     billingAddress: billingAddress;
-//     billingAddress_text: string;
-//     class: classification;
-//     createdDate: string;
-//     creditLimit: number;
-//     currency: currency;
-//     customForm: NS_Reference;
-//     department: department;
-//     discountAmount: number;
-//     discountDate: string;
-//     documentStatus: string;
-//     dueDate: string;
-//     endDate: string;
-//     entity: 'customer' | 'partner' | 'vendor' | 'nsResource' | 'employee' | 'contact';
-//     entityTaxRegNum: string;
-//     exchangeRate: number;
-//     excludeFromGLNumbering: boolean;
-//     expense: expense[];
-//     externalId: string;
-//     id: string;
-//     installment: installment[];
-//     intercoStatus: NS_Reference;
-//     intercoTransaction: nsResource;
-//     item: item[];
-//     landedCostMethod: NS_Reference;
-//     lastModifiedDate: string;
-//     links: NS_Link[];
-//     location: location;
-//     memo: string;
-//     nextApprover: employee;
-//     nexus: nexus;
-//     overrideInstallments: boolean;
-//     paymentHold: boolean;
-//     postingPeriod: accountingPeriod;
-//     prevDate: string;
-//     received: boolean;
-//     refName: string;
-//     startDate: string;
-//     status: NS_Reference;
-//     subsidiary: subsidiary;
-//     subsidiaryTaxRegNum: string;
-//     taxDetails: taxDetails[];
-//     taxDetailsOverride: boolean;
-//     taxPointDate: string;
-//     taxPointDateOverride: boolean;
-//     taxRegOverride: boolean;
-//     taxTotal: number;
-//     terms: term;
-//     toBePrinted: boolean;
-//     total: number;
-//     totalAfterTaxes: number;
-//     tranDate: string;
-//     tranId: string;
-//     unbilledOrders: number;
-//     userTotal: number;
-// }
-
-// export interface accountingBookDetail {
-//     items: accountingBookDetailElement;
-// }
-
-// export interface accountingBookDetailElement {
-//     accountingBook: accountingBook;
-//     exchangeRate: number;
-//     links: NS_Link[];
-//     refName: string;
-//     revRecOnRevCommitment: boolean;
-//     subsidiary: subsidiary;
-//     tranIsVsoeBundle: boolean;
-// }
-
-// export interface appliedRules {
-//     items: appliedRulesElement;
-// }
-
-// export interface appliedRulesElement {
-//     creationDate: string;
-//     details: string;
-//     externalLogId: number;
-//     id: number;
-//     links: NS_Link[];
-//     refName: string;
-//     ruleType: string;
-//     ruleTypeTranslation: string;
-//     transactionVersion: number;
-// }
-
-// export interface billingAddress {
-//     addr1: string;
-//     addr2: string;
-//     addr3: string;
-//     addrPhone: string;
-//     addrText: string;
-//     addressee: string;
-//     attention: string;
-//     city: string;
-//     country: NS_Reference;
-//     externalId: string;
-//     lastModifiedDate: string;
-//     links: NS_Link[];
-//     override: boolean;
-//     refName: string;
-//     state: string;
-//     zip: string;
-// }
-
-// export interface expense {
-//     items: expenseElement;
-// }
-
-// export interface expenseElement {
-//     account: account;
-//     amortizStartDate: string;
-//     amortizationEndDate: string;
-//     amortizationResidual: string;
-//     amortizationSched: amortizationSchedule;
-//     amortizationType: string;
-//     amount: number;
-//     category: expenseCategory;
-//     class: classification;
-//     customer: [customer, partner, vendor, nsResource, employee, contact];
-//     department: department;
-//     isBillable: boolean;
-//     line: number;
-//     links: NS_Link[];
-//     location: location;
-//     memo: string;
-//     orderDoc: string;
-//     orderLine: string;
-//     projecttask: projectTask;
-//     refName: string;
-//     scheduleType: string;
-//     taxAmount: number;
-//     taxDetailsReference: string;
-// }
-
-// export interface installment {
-//     items: installmentElement;
-// }
-
-// export interface installmentElement {
-//     amount: number;
-//     amountDue: number;
-//     dueDate: string;
-//     links: NS_Link[];
-//     refName: string;
-//     seqNum: number;
-// }
-
-// export interface itemInventoryDetail {
-//     customForm: NS_Reference;
-//     externalId: string;
-//     inventoryassignment: inventoryassignmentElement[];
-//     item: [
-//         inventoryItem,
-//         assemblyItem,
-//         kitItem,
-//         nsResource,
-//         discountItem,
-//         markupItem,
-//         subtotalItem,
-//         descriptionItem,
-//         paymentItem,
-//         salesTaxItem,
-//         taxGroup,
-//         shipItem,
-//         downloadItem,
-//         giftCertificateItem,
-//         subscriptionPlan,
-//         nonInventorySaleItem,
-//         nonInventoryResaleItem,
-//         nonInventoryPurchaseItem,
-//         serviceSaleItem,
-//         serviceResaleItem,
-//         servicePurchaseItem,
-//         otherChargeSaleItem,
-//         otherChargeResaleItem,
-//         otherChargePurchaseItem
-//     ];
-//     itemDescription: string;
-//     links: NS_Link[];
-//     location: location;
-//     quantity: number;
-//     refName: string;
-//     toLocation: location;
-//     unit: string;
-// }
-
-// export interface inventoryassignmentElement {
-//     binNumber: bin;
-//     expirationDate: string;
-//     internalId: number;
-//     inventoryDetail: number;
-//     inventoryStatus: inventoryStatus;
-//     issueInventoryNumber: inventoryNumber;
-//     links: NS_Link[];
-//     packCarton: string;
-//     pickCarton: string;
-//     quantity: number;
-//     quantityAvailable: number;
-//     receiptInventoryNumber: string;
-//     refName: string;
-//     secondaryQuantity: number;
-//     toBinNumber: bin;
-//     toInventoryStatus: inventoryStatus;
-// }
-
-// export interface itemElement {
-//     amortizStartDate: string;
-//     amortizationEndDate: string;
-//     amortizationResidual: string;
-//     amortizationSched: amortizationSchedule;
-//     amortizationType: string;
-//     amount: number;
-//     billReceipts: itemReceipt[];
-//     billVarianceStatus: NS_Reference;
-//     catchUpPeriod: accountingPeriod;
-//     class: classification;
-//     customer: [customer, partner, vendor, nsResource, employee, contact];
-//     deferRevRec: boolean;
-//     department: department;
-//     description: string;
-//     dueToFromSubsidiary: subsidiary;
-//     initOqpBucket: string;
-//     inventoryDetail: itemInventoryDetail;
-//     isBillable: boolean;
-//     isCatchWeightItem: boolean;
-//     isClosed: boolean;
-//     isOpen: boolean;
-//     isVsoeBundle: string;
-//     item: [inventoryItem, assemblyItem, kitItem, nsResource, discountItem, markupItem, subtotalItem, descriptionItem, paymentItem, salesTaxItem, taxGroup, shipItem, downloadItem, giftCertificateItem, subscriptionPlan, nonInventorySaleItem, nonInventoryResaleItem, nonInventoryPurchaseItem, serviceSaleItem, serviceResaleItem, servicePurchaseItem, otherChargeSaleItem, otherChargeResaleItem, otherChargePurchaseItem];
-//     itemSubtype: NS_Reference;
-//     itemType: NS_Reference;
-//     job: job;
-//     landedCostCategory: costCategory;
-//     licenseCode: string;
-//     line: number;
-//     linked: boolean;
-//     links: NS_Link[];
-//     location: location;
-//     marginal: boolean;
-//     matrixType: string;
-//     minQty: number;
-//     options: string;
-//     oqpBucket: string;
-//     orderDoc: nsResource;
-//     orderLine: number;
-//     primaryToSecondaryUnitConversionRate: number;
-//     printItems: boolean;
-//     projecttask: projectTask;
-//     quantity: number;
-//     rate: number;
-//     rateIncludingTax: number;
-//     rateSchedule: string;
-//     refName: string;
-//     scheduleType: string;
-//     secondaryQuantity: number;
-//     secondaryUnitConversionRate: number;
-//     secondaryUnits: string;
-//     secondaryUnitsList: string;
-//     subsidiary: subsidiary;
-//     targetLocation: location;
-//     taxAmount: number;
-//     taxDetailsReference: string;
-//     units: string;
-//     vendorName: string;
-// }
-
-// export interface taxDetails {
-//     items: taxDetailsElement;
-// }
-
-// export interface taxDetailsElement {
-//     calcDetail: string;
-//     lineName: string;
-//     lineType: string;
-//     links: NS_Link[];
-//     netAmount: number;
-//     refName: string;
-//     taxAmount: number;
-//     taxBasis: number;
-//     taxCode: salesTaxItem;
-//     taxDetailsReference: NS_Reference;
-//     taxRate: number;
-//     taxType: string;
-// }
-
-//////////////////////////
-/////////////////////
-////////////////
 
 //  Vendor Bill
 export interface NS_VendorBillLine {
-    // Required fields
+    // Required
     amount?: number;
     item?: {
         id: string;
